+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v1.0.0] - 2022-09-13
+
+Finally ready for a beta release!
 
 ### Added
 
-* Pulsers and hangar icons added.
+* Remaining weapons and systems added (except for gunboats).
 
 ### Changed